+++ src/Traits/WithColumnSelect.php
@@ -27,13 +27,7 @@ trait WithColumnSelect
         }
 
         // Get a list of visible default columns that are not excluded
-        $columns = collect($this->getColumns())
-            ->filter(function ($column) {
-                return $column->isVisible() && $column->isSelectable() && $column->isSelected();
-            })
-            ->map(fn ($column) => $column->getSlug())
-            ->values()
-            ->toArray();
+        $columns = $this->getDefaultVisibleColumns();
 
         // Set to either the default set or what is stored in the session
         $this->selectedColumns = (isset($this->{$this->tableName}['columns']) && count($this->{$this->tableName}['columns']) > 0) ?
@@ -49,9 +43,38 @@ trait WithColumnSelect
         }
     }
 
+    public function getDefaultVisibleColumns(): array
+    {
+        return collect($this->getColumns())
+        ->filter(function ($column) {
+            return $column->isVisible() && $column->isSelectable() && $column->isSelected();
+        })
+        ->map(fn ($column) => $column->getSlug())
+        ->values()
+        ->toArray();
+    }
+
+    public function selectAllColumns()
+    {
+        $this->{$this->tableName}['columns'] = [];
+        $this->forgetColumnSelectSession();
+    }
+
+    public function deselectAllColumns()
+    {
+        $this->{$this->tableName}['columns'] = [];
+        $this->selectedColumns = [];
+        session([$this->getColumnSelectSessionKey() => []]);
+    }
+
     public function updatedSelectedColumns(): void
     {
-        $this->{$this->tableName}['columns'] = $this->selectedColumns;
-        session([$this->getColumnSelectSessionKey() => $this->{$this->tableName}['columns']]);
+        // The query string isn't needed if it's the same as the default
+        if(count($this->selectedColumns) === count($this->getDefaultVisibleColumns())){
+            $this->selectAllColumns();
+        }else{
+            $this->{$this->tableName}['columns'] = $this->selectedColumns;
+            session([$this->getColumnSelectSessionKey() => $this->{$this->tableName}['columns']]);
+        }
     }
 }