+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 
 ## [Unreleased]
 
-## [dev-column-select] - 2022-05-14
+## [column-select-dev] - 2022-05-14
+
+### Added
 
 - Added functionality to bookmark or deep link column selection
 - Added functionality to identify different datatable components as unique in column selection and filter arrays
+
+### Changed
+
+- Simplified misc/multiple-tables.md in docs where additional steps are no longer needed. (Original steps still work and lead to same result)
 
 ## [2.7.0] - 2022-05-07